+++ worlds/factorio_space_age/CHANGELOG.md
@@ -11,6 +11,8 @@
 * New YAML option `production_and_utility_science: removed` disabled by default, which allows for a smaller scale game, particularly with shorter objectives like `goal: any_other_planet_science`.
 * Research objectives beyond your configured goal are removed. For example, `goal: aquilo_orbit` will remove all research locations involving cryogenic science packs. When combined with `infinite_technologies: shuffle` (the default), infinite technologies beyond the goal are moved earlier instead of removed, copying the ingredients, prerequisites, and count formula of an earlier infinite technology. This can enable Factorio: Space Age players to unblock themselves by researching large progressive groups from infinite technology locations out of logic once enough science packs are unlocked, always at least chemical science. [#11](https://github.com/thejoshwolfe/Archipelago/issues/11)
 * Fix filler items not contributing to their progressive group like they're supposed to.
+* Improved information in the technology GUI:
+    * Progressive chains are shown as dependencies between the inaccessible vanilla technologies.
 
 ## 2.2.4